bootrom-tools: bin2verilog: adapted to handle odd-length files
The output format is in DWORDS, and the tool was originally written
assuming that the input bin file would similarly be some multiple of DWORDs
long. This proved to be not the case, but hadn't shown up in testing. It now
zero-pads the input buffer to be suitably sized.

Testing Done: ran it against good and previously-failing .bin files
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by: Morgan Girling <morgang@bsquare.com>