bootrom-tools: removed moribund code
Initial bootrom tool (create-tftf) was written in C. As a result of review
comments, that tool and all subsequent tools done in python under
bootrom-tools/scripts. The C folders were retained during the transition
to python as they provided some useful bits and pieces, but the code is
no longer needed and its presence confused people.

This removes the original C, include and test folders, as well as the
makefile, which is no longer needed. As the tree now looked like a
Charlie Brown Christmas tree, I moved the content of the scripts folder
up to the bootrom-tools root, and updated README.md to reflect the move.

Testing Done: Verified that scripts continue to work
Bug Number: trivial
Review URL:
Reviewed by: Greg KH, Eli Sennesh

Signed-off-by: Morgan Girling <morgang@bsquare.com>