Bootrom tools: test automation: added "--efuse" as a per-test option
Run-bootrom-tests previously took an e-Fuse file parameter, which was applied
globally to all tests. To allow for testing the e-Fuse error detection code,
we needed to be able to inject bogus e-Fuse values in specific tests. This is
done by adding the --efuse option to the test line to specify an alternate
e-Fuse file to use on that test. Support for this necessarily percolated up to
the create-bootrom-test-suite app, which generates test scripts.

Testing Done: Dev. testing
Bug Number: SW-982
Review URL: https://reviews.leaflabs.com/r/1474/
Reviewed by: Jun Li

Signed-off-by:  Morgan Girling <morgang@bsquare.com>