Bootrom tools: Refined error handling for mysterious error
I've been more frequently running into an unexplained "Downloading file
[<filename>]...Unspecified error -1" from the download monitor. This
proved to be due to an intermittent USB connection to the J-Link JTAG
adapter. This change checks for the "Could not find emulator with USB
serial number" and prints out a probable cause for the mysterious error.

Testing Done: Ran the code with the j-Link connected and disconnected, and
noted the additional error message when disconnected.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1561/
Reviewed by: Jun Li, Eli Sennesh

Signed-off-by:  Morgan Girling <morgang@bsquare.com>