Bootrom Tools: Revised test data
With the changes to the error logging and DME-write logging, revamped a
subset of tests to reflect the new responses. This includes changes to the
.tss (test script script) file. Addition of the eccerror test required
additions of the eccerror register to existing tools. Also renamed the
tests and the response files.

Testing Done: Ran selected tests and harvested the results
Bug Number: trivial
Review URL:
Reviewed by: Marti Bolivar

Signed-off-by:  Morgan Girling <morgang@bsquare.com>