Force not to build second stage FW
Our ultimate design is to have boot ROM load second stage FW, then
it would load the third stage FW.
For simulation done in Toshiba, this is not used. We need to skip
the second stage FW and package the third stage FW as the image
loaded directly by boot ROM.
Current script has parameter to do that but looks like it is very
easy to be forgotten when making the drop.
And there is no reason we can think of at this moment why we want
to include the second stage FW in the build for Toshiba.
So just force it to be the case all the time.

Testing Done:   The second stage FW is skipped.
Bug Number:     Trivial
Review URL:     https://reviews.leaflabs.com/r/1807/
Reviewed by:    Olin Sibert <osibert@oxford.com>
                Morgan Girling <girling_morgan@projectara.com>

Signed-off-by:  Jun Li <li_jun@projectara.com>