Bootrom Tools: Build Tools: Expose the DME logging as a build flag
Previously, DME logging to dbgserial was done on an ad hoc basis by enabling
the #define locally, but not checking it in. To better intergrate with the
drop tools, etc. this is now exposed through the "-dme" option in all of the
build tools.

Testing Done: verified transport via make3
Bug Number: SW-1205
Review URL: https://reviews.leaflabs.com/r/1835/
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>