Bootrom Tools: Test Tools: Test a log file against a response file
Compare a log file against a response file. The response file may be a subset
of the log file - the intent is for it to consist of salient (sub)strings to
look for in the log file. The strings must all be encountered in the order
they are specified in the response file for the test to be considered
successful. The app. exits with a zero status if the test passed. If the test
failed it will issue an error message indicating the first missed line in the
response file and exit with a non-zero status of 2.

Testing Done: compared against logs with variants of the response file to test
success and failure cases.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1845/
Reviewed by:Olin Sibert, Jun Li

Signed-off-by:  Morgan Girling <morgang@bsquare.com>